refactor: mv request mod to crate level
